pyproject.toml
[project]
name = "form-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "ipdb>=0.13.13",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.12.0",
    "pytest-selenium>=4.1.0",
]
